+++ src/rsi_calc.cpp
@@ -1,12 +1,11 @@
 #include <Rcpp.h>
-#include <vector>        // for std::vector
 #include <functional>    // for std::less, etc
 #include <algorithm>     // for count_if
 
-using namespace Rcpp ;
+using namespace Rcpp;
 
 // [[Rcpp::export]]
-int rsi_calc_S(std::vector<double> x, bool include_I) {
+int rsi_calc_S(DoubleVector x, bool include_I) {
   if (include_I == TRUE) {
     return count_if(x.begin(), x.end(), bind2nd(std::less_equal<double>(), 2));
   } else {
@@ -15,7 +14,7 @@ int rsi_calc_S(std::vector<double> x, bool include_I) {
 }
 
 // [[Rcpp::export]]
-int rsi_calc_R(std::vector<double> x, bool include_I) {
+int rsi_calc_R(DoubleVector x, bool include_I) {
   if (include_I == TRUE) {
     return count_if(x.begin(), x.end(), bind2nd(std::greater_equal<double>(), 2));
   } else {
@@ -24,6 +23,6 @@ int rsi_calc_R(std::vector<double> x, bool include_I) {
 }
 
 // [[Rcpp::export]]
-int rsi_calc_total(std::vector<double> x) {
- return count_if(x.begin(), x.end(), bind2nd(std::less_equal<double>(), 3));
+int rsi_calc_total(DoubleVector x) {
+  return count_if(x.begin(), x.end(), bind2nd(std::less_equal<double>(), 3));
 }